small code cleanup changes
